\name{covariance-package}
\alias{covariance-package}
\alias{covariance}
\docType{package}
\title{
Methods to estimate covariance matrices
}
\description{
Estimate covariance matrices using robust, shrinkage, and other estimates
}
\details{
\tabular{ll}{
Package: \tab covariance\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-12-18\cr
License: \tab What license is it under?\cr
}
TODO
}
\author{
Ross Bennett

Maintainer: Ross Bennett <rossbennett34@gmail.com>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
